6 Down Close
"I took the precious case into my hands and looked in some perplexity from it to my illustrious
client.

"'You doubt its value?' he asked.

"'Not at all. I only doubt--'

"'The propriety of my leaving it. You may set your mind at rest about that. I should not dream of
doing so were it not absolutely certain that I should be able in four days to reclaim it. It is a
pure matter of form. Is the security sufficient?'

"'Ample.'

"'You understand, Mr. Holder, that I am giving you a strong proof of the confidence which I have in
you, founded upon all that I have heard of you. I rely upon you not only to be discreet and to
refrain from all gossip upon the matter but, above all, to preserve this coronet with every possible
precaution because I need not say that a great public scandal would be caused if any harm were to
befall it. Any injury to it would be almost as serious as its complete loss, for there are no beryls
in the world to match these, and it would be impossible to replace them. I leave it with you,
however, with every confidence, and I shall call for it in person on Monday morning.'