140 Wantage Close
"'I should be happy to advance it without further parley from my own private purse,' said I, 'were
it not that the strain would be rather more than it could bear. If, on the other hand, I am to do it
in the name of the firm, then in justice to my partner I must insist that, even in your case, every
businesslike precaution should be taken.'

"'I should much prefer to have it so,' said he, raising up a square, black morocco case which he had
laid beside his chair.  'You have doubtless heard of the Beryl Coronet?'

"'One of the most precious public possessions of the empire,' said I.

"'Precisely.' He opened the case, and there, imbedded in soft, flesh-coloured velvet, lay the
magnificent piece of jewellery which he had named. 'There are thirty-nine enormous beryls,' said he,
'and the price of the gold chasing is incalculable. The lowest estimate would put the worth of the
coronet at double the sum which I have asked. I am prepared to leave it with you as my security.'