122 Balcombe Close
I had no keener pleasure than in following Holmes in his professional investigations, and in
admiring the rapid deductions, as swift as intuitions, and yet always founded on a logical basis
with which he unravelled the problems which were submitted to him. I rapidly threw on my clothes and
was ready in a few minutes to accompany my friend down to the sitting-room. A lady dressed in black
and heavily veiled, who had been sitting in the window, rose as we entered.

"Good-morning, madam," said Holmes cheerily. "My name is Sherlock Holmes. This is my intimate friend
and associate, Dr. Watson, before whom you can speak as freely as before myself. Ha! I am glad to
see that Mrs. Hudson has had the good sense to light the fire. Pray draw up to it, and I shall order
you a cup of hot coffee, for I observe that you are shivering."

"It is not cold which makes me shiver," said the woman in a low voice, changing her seat as
requested.

"What, then?"

"It is fear, Mr. Holmes. It is terror." She raised her veil as she spoke, and we could see that she
was indeed in a pitiable state of agitation, her face all drawn and grey, with restless frightened
eyes, like those of some hunted animal. Her features and figure were those of a woman of thirty, but
her hair was shot with premature grey, and her expression was weary and haggard. Sherlock Holmes ran
her over with one of his quick, all-comprehensive glances.

"You must not fear," said he soothingly, bending forward and patting her forearm. "We shall soon set
matters right, I have no doubt. You have come in by train this morning, I see."