160 Down Close
"Yes; when I went to see if the door was fastened for the night I met her slipping in. I saw the
man, too, in the gloom."

"Do you know him?"

"Oh, yes! he is the green-grocer who brings our vegetables round.  His name is Francis Prosper."

"He stood," said Holmes, "to the left of the door--that is to say, farther up the path than is
necessary to reach the door?"