15 Ventnor Terrace
"You are sure that she has not sent it yet?"

"I am sure."

"And why?"

"Because she has said that she would send it on the day when the betrothal was publicly proclaimed.
That will be next Monday."

"Oh, then we have three days yet," said Holmes with a yawn. "That is very fortunate, as I have one
or two matters of importance to look into just at present. Your Majesty will, of course, stay in
London for the present?"

"Certainly. You will find me at the Langham under the name of the Count Von Kramm."