77 Larch Walk
"'Thank you, Maggie,' says I; 'but if it is all the same to you, I'd rather have that one I was
handling just now.'

"'The other is a good three pound heavier,' said she, 'and we fattened it expressly for you.'

"'Never mind. I'll have the other, and I'll take it now,' said I.

"'Oh, just as you like,' said she, a little huffed. 'Which is it you want, then?'