23 Glendale
Sherlock Holmes was not very communicative during the long drive and lay back in the cab humming the
tunes which he had heard in the afternoon. We rattled through an endless labyrinth of gas-lit
streets until we emerged into Farrington Street.

"We are close there now," my friend remarked. "This fellow Merryweather is a bank director, and
personally interested in the matter. I thought it as well to have Jones with us also. He is not a
bad fellow, though an absolute imbecile in his profession. He has one positive virtue. He is as
brave as a bulldog and as tenacious as a lobster if he gets his claws upon anyone. Here we are, and
they are waiting for us."