105 Wantage Close
"This discovery, and the evident confusion which the cripple showed, made the inspector realise that
the matter was serious. The rooms were carefully examined, and results all pointed to an abominable
crime. The front room was plainly furnished as a sitting-room and led into a small bedroom, which
looked out upon the back of one of the wharves. Between the wharf and the bedroom window is a narrow
strip, which is dry at low tide but is covered at high tide with at least four and a half feet of
water. The bedroom window was a broad one and opened from below. On examination traces of blood were
to be seen upon the windowsill, and several scattered drops were visible upon the wooden floor of
the bedroom. Thrust away behind a curtain in the front room were all the clothes of Mr. Neville St.
Clair, with the exception of his coat. His boots, his socks, his hat, and his watch--all were there.
There were no signs of violence upon any of these garments, and there were no other traces of Mr.
Neville St. Clair. Out of the window he must apparently have gone for no other exit could be
discovered, and the ominous bloodstains upon the sill gave little promise that he could save himself
by swimming, for the tide was at its very highest at the moment of the tragedy.

"And now as to the villains who seemed to be immediately implicated in the matter. The Lascar was
known to be a man of the vilest antecedents, but as, by Mrs. St. Clair's story, he was known to have
been at the foot of the stair within a very few seconds of her husband's appearance at the window,
he could hardly have been more than an accessory to the crime. His defence was one of absolute
ignorance, and he protested that he had no knowledge as to the doings of Hugh Boone, his lodger, and
that he could not account in any way for the presence of the missing gentleman's clothes.