142 Moor End
"A diamond, sir? A precious stone. It cuts into glass as though it were putty."

"It's more than a precious stone. It is the precious stone."

"Not the Countess of Morcar's blue carbuncle!" I ejaculated.

"Precisely so. I ought to know its size and shape, seeing that I have read the advertisement about
it in The Times every day lately. It is absolutely unique, and its value can only be conjectured,
but the reward offered of £1000 is certainly not within a twentieth part of the market price."

"A thousand pounds! Great Lord of mercy!" The commissionaire plumped down into a chair and stared
from one to the other of us.