79 Glan Road
He held up a piece of white cardboard about the size of a sheet of note-paper. It read in this
fashion:

The Red-headed League          is       Dissolved   October 9, 1890.

Sherlock Holmes and I surveyed this curt announcement and the rueful face behind it, until the
comical side of the affair so completely overtopped every other consideration that we both burst out
into a roar of laughter.