141 Glan Road
"I have seen too much not to know that the impression of a woman may be more valuable than the
conclusion of an analytical reasoner. And in this letter you certainly have a very strong piece of
evidence to corroborate your view. But if your husband is alive and able to write letters, why
should he remain away from you?"

"I cannot imagine. It is unthinkable."

"And on Monday he made no remarks before leaving you?"

"No."

"And you were surprised to see him in Swandam Lane?"