58 Cliff Place
"The paper was made in Bohemia," I said.

"Precisely. And the man who wrote the note is a German. Do you note the peculiar construction of the
sentence--'This account of you we have from all quarters received.' A Frenchman or Russian could not
have written that. It is the German who is so uncourteous to his verbs. It only remains, therefore,
to discover what is wanted by this German who writes upon Bohemian paper and prefers wearing a mask
to showing his face. And here he comes, if I am not mistaken, to resolve all our doubts."

As he spoke there was the sharp sound of horses' hoofs and grating wheels against the curb, followed
by a sharp pull at the bell. Holmes whistled.

"A pair, by the sound," said he. "Yes," he continued, glancing out of the window. "A nice little
brougham and a pair of beauties. A hundred and fifty guineas apiece. There's money in this case,
Watson, if there is nothing else."

"I think that I had better go, Holmes."