105 Badgers Dene
"You can at least tell me whether my own thought is correct, and if she died from some sudden
fright."

"No, I do not think so. I think that there was probably some more tangible cause. And now, Miss
Stoner, we must leave you for if Dr. Roylott returned and saw us our journey would be in vain.
Good-bye, and be brave, for if you will do what I have told you, you may rest assured that we shall
soon drive away the dangers that threaten you."

Sherlock Holmes and I had no difficulty in engaging a bedroom and sitting-room at the Crown Inn.
They were on the upper floor, and from our window we could command a view of the avenue gate, and of
the inhabited wing of Stoke Moran Manor House. At dusk we saw Dr. Grimesby Roylott drive past, his
huge form looming up beside the little figure of the lad who drove him. The boy had some slight
difficulty in undoing the heavy iron gates, and we heard the hoarse roar of the doctor's voice and
saw the fury with which he shook his clinched fists at him. The trap drove on, and a few minutes
later we saw a sudden light spring up among the trees as the lamp was lit in one of the sitting-
rooms.