50 Harvey Street
"'Oh, just as you like,' said she, a little huffed. 'Which is it you want, then?'

"'That white one with the barred tail, right in the middle of the flock.'

"'Oh, very well. Kill it and take it with you.'

"Well, I did what she said, Mr. Holmes, and I carried the bird all the way to Kilburn. I told my pal
what I had done, for he was a man that it was easy to tell a thing like that to. He laughed until he
choked, and we got a knife and opened the goose. My heart turned to water, for there was no sign of
the stone, and I knew that some terrible mistake had occurred. I left the bird, rushed back to my
sister's, and hurried into the back yard. There was not a bird to be seen there.