107 Valentia Road
"Ah, yes. I should like a few particulars as to this young lady, and your relations to her."

Lord St. Simon shrugged his shoulders and raised his eyebrows. "We have been on a friendly footing
for some years--I may say on a very friendly footing. She used to be at the Allegro. I have not
treated her ungenerously, and she had no just cause of complaint against me, but you know what women
are, Mr. Holmes. Flora was a dear little thing, but exceedingly hot-headed and devotedly attached to
me. She wrote me dreadful letters when she heard that I was about to be married, and, to tell the
truth, the reason why I had the marriage celebrated so quietly was that I feared lest there might be
a scandal in the church. She came to Mr. Doran's door just after we returned, and she endeavoured to
push her way in, uttering very abusive expressions towards my wife, and even threatening her, but I
had foreseen the possibility of something of the sort, and I had two police fellows there in private
clothes, who soon pushed her out again. She was quiet when she saw that there was no good in making
a row."

"Did your wife hear all this?"

"No, thank goodness, she did not."

"And she was seen walking with this very woman afterwards?"

"Yes. That is what Mr. Lestrade, of Scotland Yard, looks upon as so serious. It is thought that
Flora decoyed my wife out and laid some terrible trap for her."