58 Down Close
"Come in!" said he.

The man who entered was young, some two-and-twenty at the outside, well-groomed and trimly clad,
with something of refinement and delicacy in his bearing. The streaming umbrella which he held in
his hand, and his long shining waterproof told of the fierce weather through which he had come. He
looked about him anxiously in the glare of the lamp, and I could see that his face was pale and his
eyes heavy, like those of a man who is weighed down with some great anxiety.