116 Sunningdale Drive
"Was she in good spirits?"

"Never better. She kept talking of what we should do in our future lives."

"Indeed! That is very interesting. And on the morning of the wedding?"

"She was as bright as possible--at least until after the ceremony."

"And did you observe any change in her then?"