84 Moor End
Our visitor staggered to his feet and clutched the mantelpiece with his right hand. Holmes unlocked
his strong-box and held up the blue carbuncle, which shone out like a star, with a cold, brilliant,
many-pointed radiance. Ryder stood glaring with a drawn face, uncertain whether to claim or to
disown it.

"The game's up, Ryder," said Holmes quietly. "Hold up, man, or you'll be into the fire! Give him an
arm back into his chair, Watson. He's not got blood enough to go in for felony with impunity. Give
him a dash of brandy. So! Now he looks a little more human. What a shrimp it is, to be sure!"

For a moment he had staggered and nearly fallen, but the brandy brought a tinge of colour into his
cheeks, and he sat staring with frightened eyes at his accuser.

"I have almost every link in my hands, and all the proofs which I could possibly need, so there is
little which you need tell me. Still, that little may as well be cleared up to make the case
complete. You had heard, Ryder, of this blue stone of the Countess of Morcar's?"