154 Wantage Close
"There's a cold-blooded scoundrel!" said Holmes, laughing, as he threw himself down into his chair
once more. "That fellow will rise from crime to crime until he does something very bad, and ends on
a gallows. The case has, in some respects, been not entirely devoid of interest."

"I cannot now entirely see all the steps of your reasoning," I remarked.

"Well, of course it was obvious from the first that this Mr. Hosmer Angel must have some strong
object for his curious conduct, and it was equally clear that the only man who really profited by
the incident, as far as we could see, was the stepfather. Then the fact that the two men were never
together, but that the one always appeared when the other was away, was suggestive. So were the
tinted spectacles and the curious voice, which both hinted at a disguise, as did the bushy whiskers.
My suspicions were all confirmed by his peculiar action in typewriting his signature, which, of
course, inferred that his handwriting was so familiar to her that she would recognise even the
smallest sample of it. You see all these isolated facts, together with many minor ones, all pointed
in the same direction."

"And how did you verify them?"