11 Wantage Close
"A proposition which I took the liberty of doubting."

"You did, Doctor, but none the less you must come round to my view, for otherwise I shall keep on
piling fact upon fact on you until your reason breaks down under them and acknowledges me to be
right. Now, Mr. Jabez Wilson here has been good enough to call upon me this morning, and to begin a
narrative which promises to be one of the most singular which I have listened to for some time. You
have heard me remark that the strangest and most unique things are very often connected not with the
larger but with the smaller crimes, and occasionally, indeed, where there is room for doubt whether
any positive crime has been committed. As far as I have heard it is impossible for me to say whether
the present case is an instance of crime or not, but the course of events is certainly among the
most singular that I have ever listened to. Perhaps, Mr. Wilson, you would have the great kindness
to recommence your narrative. I ask you not merely because my friend Dr. Watson has not heard the
opening part but also because the peculiar nature of the story makes me anxious to have every
possible detail from your lips. As a rule, when I have heard some slight indication of the course of
events, I am able to guide myself by the thousands of other similar cases which occur to my memory.
In the present instance I am forced to admit that the facts are, to the best of my belief, unique."

The portly client puffed out his chest with an appearance of some little pride and pulled a dirty
and wrinkled newspaper from the inside pocket of his greatcoat. As he glanced down the advertisement
column, with his head thrust forward and the paper flattened out upon his knee, I took a good look
at the man and endeavoured, after the fashion of my companion, to read the indications which might
be presented by his dress or appearance.

I did not gain very much, however, by my inspection. Our visitor bore every mark of being an average
commonplace British tradesman, obese, pompous, and slow. He wore rather baggy grey shepherd's check
trousers, a not over-clean black frock-coat, unbuttoned in the front, and a drab waistcoat with a
heavy brassy Albert chain, and a square pierced bit of metal dangling down as an ornament. A frayed
top-hat and a faded brown overcoat with a wrinkled velvet collar lay upon a chair beside him.
Altogether, look as I would, there was nothing remarkable about the man save his blazing red head,
and the expression of extreme chagrin and discontent upon his features.

Sherlock Holmes' quick eye took in my occupation, and he shook his head with a smile as he noticed
my questioning glances. "Beyond the obvious facts that he has at some time done manual labour, that
he takes snuff, that he is a Freemason, that he has been in China, and that he has done a
considerable amount of writing lately, I can deduce nothing else."

Mr. Jabez Wilson started up in his chair, with his forefinger upon the paper, but his eyes upon my
companion.