195 Wantage Close
"I should like to see him very much."

"Would you? That is easily done. Come this way. You can leave your bag."

"No, I think that I'll take it."

"Very good. Come this way, if you please." He led us down a passage, opened a barred door, passed
down a winding stair, and brought us to a whitewashed corridor with a line of doors on each side.

"The third on the right is his," said the inspector. "Here it is!" He quietly shot back a panel in
the upper part of the door and glanced through.