10 Sunningdale Drive
"This was quite too good to lose, Watson. I was just balancing whether I should run for it, or
whether I should perch behind her landau when a cab came through the street. The driver looked twice
at such a shabby fare, but I jumped in before he could object. 'The Church of St. Monica,' said I,
'and half a sovereign if you reach it in twenty minutes.' It was twenty-five minutes to twelve, and
of course it was clear enough what was in the wind.

"My cabby drove fast. I don't think I ever drove faster, but the others were there before us. The
cab and the landau with their steaming horses were in front of the door when I arrived. I paid the
man and hurried into the church. There was not a soul there save the two whom I had followed and a
surpliced clergyman, who seemed to be expostulating with them. They were all three standing in a
knot in front of the altar. I lounged up the side aisle like any other idler who has dropped into a
church. Suddenly, to my surprise, the three at the altar faced round to me, and Godfrey Norton came
running as hard as he could towards me.