30 Moor End
It seems that it was. She had the surest information that of late he had, when the fit was on him,
made use of an opium den in the farthest east of the City. Hitherto his orgies had always been
confined to one day, and he had come back, twitching and shattered, in the evening. But now the
spell had been upon him eight-and-forty hours, and he lay there, doubtless among the dregs of the
docks, breathing in the poison or sleeping off the effects. There he was to be found, she was sure
of it, at the Bar of Gold, in Upper Swandam Lane. But what was she to do? How could she, a young and
timid woman, make her way into such a place and pluck her husband out from among the ruffians who
surrounded him?

There was the case, and of course there was but one way out of it. Might I not escort her to this
place? And then, as a second thought, why should she come at all? I was Isa Whitney's medical
adviser, and as such I had influence over him. I could manage it better if I were alone. I promised
her on my word that I would send him home in a cab within two hours if he were indeed at the address
which she had given me. And so in ten minutes I had left my armchair and cheery sitting-room behind
me, and was speeding eastward in a hansom on a strange errand, as it seemed to me at the time,
though the future only could show how strange it was to be.

But there was no great difficulty in the first stage of my adventure. Upper Swandam Lane is a vile
alley lurking behind the high wharves which line the north side of the river to the east of London
Bridge. Between a slop-shop and a gin-shop, approached by a steep flight of steps leading down to a
black gap like the mouth of a cave, I found the den of which I was in search. Ordering my cab to
wait, I passed down the steps, worn hollow in the centre by the ceaseless tread of drunken feet; and
by the light of a flickering oil-lamp above the door I found the latch and made my way into a long,
low room, thick and heavy with the brown opium smoke, and terraced with wooden berths, like the
forecastle of an emigrant ship.