48 Ventnor Terrace
"I have said that though the floor and ceiling were of iron, the walls were of wood. As I gave a
last hurried glance around, I saw a thin line of yellow light between two of the boards, which
broadened and broadened as a small panel was pushed backward. For an instant I could hardly believe
that here was indeed a door which led away from death. The next instant I threw myself through, and
lay half-fainting upon the other side. The panel had closed again behind me, but the crash of the
lamp, and a few moments afterwards the clang of the two slabs of metal, told me how narrow had been
my escape.

"I was recalled to myself by a frantic plucking at my wrist, and I found myself lying upon the stone
floor of a narrow corridor, while a woman bent over me and tugged at me with her left hand, while
she held a candle in her right. It was the same good friend whose warning I had so foolishly
rejected.