46 Cliff Place
"That is our signal," said Holmes, springing to his feet; "it comes from the middle window."

As we passed out he exchanged a few words with the landlord, explaining that we were going on a late
visit to an acquaintance, and that it was possible that we might spend the night there. A moment
later we were out on the dark road, a chill wind blowing in our faces, and one yellow light
twinkling in front of us through the gloom to guide us on our sombre errand.