103 Beaconside
"'That is quite correct,' I answered; 'but you will excuse me if I say that I cannot see how all
this bears upon my professional qualifications. I understand that it was on a professional matter
that you wished to speak to me?'

"'Undoubtedly so. But you will find that all I say is really to the point. I have a professional
commission for you, but absolute secrecy is quite essential--absolute secrecy, you understand, and
of course we may expect that more from a man who is alone than from one who lives in the bosom of
his family.'

"'If I promise to keep a secret,' said I, 'you may absolutely depend upon my doing so.'

"He looked very hard at me as I spoke, and it seemed to me that I had never seen so suspicious and
questioning an eye.

"'Do you promise, then?' said he at last.