166 Wantage Close
"My cabby drove fast. I don't think I ever drove faster, but the others were there before us. The
cab and the landau with their steaming horses were in front of the door when I arrived. I paid the
man and hurried into the church. There was not a soul there save the two whom I had followed and a
surpliced clergyman, who seemed to be expostulating with them. They were all three standing in a
knot in front of the altar. I lounged up the side aisle like any other idler who has dropped into a
church. Suddenly, to my surprise, the three at the altar faced round to me, and Godfrey Norton came
running as hard as he could towards me.

"'Thank God,' he cried. 'You'll do. Come! Come!'

"'What then?' I asked.

"'Come, man, come, only three minutes, or it won't be legal.'

"I was half-dragged up to the altar, and before I knew where I was I found myself mumbling responses
which were whispered in my ear, and vouching for things of which I knew nothing, and generally
assisting in the secure tying up of Irene Adler, spinster, to Godfrey Norton, bachelor. It was all
done in an instant, and there was the gentleman thanking me on the one side and the lady on the
other, while the clergyman beamed on me in front. It was the most preposterous position in which I
ever found myself in my life, and it was the thought of it that started me laughing just now. It
seems that there had been some informality about their license, that the clergyman absolutely
refused to marry them without a witness of some sort, and that my lucky appearance saved the
bridegroom from having to sally out into the streets in search of a best man. The bride gave me a
sovereign, and I mean to wear it on my watch-chain in memory of the occasion."

"This is a very unexpected turn of affairs," said I; "and what then?"