47 Cliff Place
"It may turn out to be of more interest than you think. You remember that the affair of the blue
carbuncle, which appeared to be a mere whim at first, developed into a serious investigation. It may
be so in this case, also."

"Well, let us hope so. But our doubts will very soon be solved, for here, unless I am much mistaken,
is the person in question."