29 Beaconside
"I didn't know what to do, so I came straight to you." That was always the way. Folk who were in
grief came to my wife like birds to a light-house.

"It was very sweet of you to come. Now, you must have some wine and water, and sit here comfortably
and tell us all about it. Or should you rather that I sent James off to bed?"