18 Wellswood Gardens
"But it is profoundly true. Singularity is almost invariably a clue. The more featureless and
commonplace a crime is, the more difficult it is to bring it home. In this case, however, they have
established a very serious case against the son of the murdered man."

"It is a murder, then?"