108 Larch Walk
"I see no marks."

"There are none."

"How do you know, then?"

"The grass was growing under it. It had only lain there a few days. There was no sign of a place
whence it had been taken. It corresponds with the injuries. There is no sign of any other weapon."

"And the murderer?"

"Is a tall man, left-handed, limps with the right leg, wears thick-soled shooting-boots and a grey
cloak, smokes Indian cigars, uses a cigar-holder, and carries a blunt pen-knife in his pocket. There
are several other indications, but these may be enough to aid us in our search."