40 Balcombe Close
"I had come to these conclusions before ever I had entered his room. An inspection of his chair
showed me that he had been in the habit of standing on it, which of course would be necessary in
order that he should reach the ventilator. The sight of the safe, the saucer of milk, and the loop
of whipcord were enough to finally dispel any doubts which may have remained. The metallic clang
heard by Miss Stoner was obviously caused by her stepfather hastily closing the door of his safe
upon its terrible occupant. Having once made up my mind, you know the steps which I took in order to
put the matter to the proof. I heard the creature hiss as I have no doubt that you did also, and I
instantly lit the light and attacked it."

"With the result of driving it through the ventilator."

"And also with the result of causing it to turn upon its master at the other side. Some of the blows
of my cane came home and roused its snakish temper, so that it flew upon the first person it saw. In
this way I am no doubt indirectly responsible for Dr. Grimesby Roylott's death, and I cannot say
that it is likely to weigh very heavily upon my conscience."

Of all the problems which have been submitted to my friend, Mr. Sherlock Holmes, for solution during
the years of our intimacy, there were only two which I was the means of introducing to his notice--
that of Mr. Hatherley's thumb, and that of Colonel Warburton's madness. Of these the latter may have
afforded a finer field for an acute and original observer, but the other was so strange in its
inception and so dramatic in its details that it may be the more worthy of being placed upon record,
even if it gave my friend fewer openings for those deductive methods of reasoning by which he
achieved such remarkable results. The story has, I believe, been told more than once in the
newspapers, but, like all such narratives, its effect is much less striking when set forth en bloc
in a single half-column of print than when the facts slowly evolve before your own eyes, and the
mystery clears gradually away as each new discovery furnishes a step which leads on to the complete
truth. At the time the circumstances made a deep impression upon me, and the lapse of two years has
hardly served to weaken the effect.