26 Corbet Close
"I came to you, sir, because I heard of you from Mrs. Etherege, whose husband you found so easy when
the police and everyone had given him up for dead. Oh, Mr. Holmes, I wish you would do as much for
me. I'm not rich, but still I have a hundred a year in my own right, besides the little that I make
by the machine, and I would give it all to know what has become of Mr. Hosmer Angel."

"Why did you come away to consult me in such a hurry?" asked Sherlock Holmes, with his finger-tips
together and his eyes to the ceiling.

Again a startled look came over the somewhat vacuous face of Miss Mary Sutherland. "Yes, I did bang
out of the house," she said, "for it made me angry to see the easy way in which Mr. Windibank--that
is, my father--took it all. He would not go to the police, and he would not go to you, and so at
last, as he would do nothing and kept on saying that there was no harm done, it made me mad, and I
just on with my things and came right away to you."