114 Cliff Place
"'Oh, very well. Kill it and take it with you.'

"Well, I did what she said, Mr. Holmes, and I carried the bird all the way to Kilburn. I told my pal
what I had done, for he was a man that it was easy to tell a thing like that to. He laughed until he
choked, and we got a knife and opened the goose. My heart turned to water, for there was no sign of
the stone, and I knew that some terrible mistake had occurred. I left the bird, rushed back to my
sister's, and hurried into the back yard. There was not a bird to be seen there.