33 Holmefield Avenue
"And how?"

"In the dress is a pocket. In the pocket is a card-case. In the card-case is a note. And here is the
very note." He slapped it down upon the table in front of him. "Listen to this:

"'You will see me when all is ready. Come at once. "'F.H.M.'