108 Bowler Street
"A considerable crime is in contemplation. I have every reason to believe that we shall be in time
to stop it. But to-day being Saturday rather complicates matters. I shall want your help to-night."

"At what time?"

"Ten will be early enough."

"I shall be at Baker Street at ten."