21 Balcombe Close
"Well, it is just as I have been telling you, Mr. Sherlock Holmes," said Jabez Wilson, mopping his
forehead; "I have a small pawnbroker's business at Coburg Square, near the City. It's not a very
large affair, and of late years it has not done more than just give me a living. I used to be able
to keep two assistants, but now I only keep one; and I would have a job to pay him but that he is
willing to come for half wages so as to learn the business."

"What is the name of this obliging youth?" asked Sherlock Holmes.

"His name is Vincent Spaulding, and he's not such a youth, either. It's hard to say his age. I
should not wish a smarter assistant, Mr. Holmes; and I know very well that he could better himself
and earn twice what I am able to give him. But, after all, if he is satisfied, why should I put
ideas in his head?"

"Why, indeed? You seem most fortunate in having an employee who comes under the full market price.
It is not a common experience among employers in this age. I don't know that your assistant is not
as remarkable as your advertisement."

"Oh, he has his faults, too," said Mr. Wilson. "Never was such a fellow for photography. Snapping
away with a camera when he ought to be improving his mind, and then diving down into the cellar like
a rabbit into its hole to develop his pictures. That is his main fault, but on the whole he's a good
worker. There's no vice in him."