86 Bowler Street
"Showing that she had struck a light and looked about her when the alarm took place. That is
important. And what conclusions did the coroner come to?"

"He investigated the case with great care, for Dr. Roylott's conduct had long been notorious in the
county, but he was unable to find any satisfactory cause of death. My evidence showed that the door
had been fastened upon the inner side, and the windows were blocked by old-fashioned shutters with
broad iron bars, which were secured every night. The walls were carefully sounded, and were shown to
be quite solid all round, and the flooring was also thoroughly examined, with the same result. The
chimney is wide, but is barred up by four large staples. It is certain, therefore, that my sister
was quite alone when she met her end. Besides, there were no marks of any violence upon her."

"How about poison?"

"The doctors examined her for it, but without success."