21 Valentia Road
"Now, Mr. Holmes, you know the people who live under my roof, and I shall continue with my miserable
story.

"When we were taking coffee in the drawing-room that night after dinner, I told Arthur and Mary my
experience, and of the precious treasure which we had under our roof, suppressing only the name of
my client. Lucy Parr, who had brought in the coffee, had, I am sure, left the room; but I cannot
swear that the door was closed. Mary and Arthur were much interested and wished to see the famous
coronet, but I thought it better not to disturb it.