60 Glan Road
"Fancy his having the insolence to confound me with the official detective force! This incident
gives zest to our investigation, however, and I only trust that our little friend will not suffer
from her imprudence in allowing this brute to trace her. And now, Watson, we shall order breakfast,
and afterwards I shall walk down to Doctors' Commons, where I hope to get some data which may help
us in this matter."

It was nearly one o'clock when Sherlock Holmes returned from his excursion. He held in his hand a
sheet of blue paper, scrawled over with notes and figures.

"I have seen the will of the deceased wife," said he. "To determine its exact meaning I have been
obliged to work out the present prices of the investments with which it is concerned. The total
income, which at the time of the wife's death was little short of £1100, is now, through the fall in
agricultural prices, not more than £750. Each daughter can claim an income of £250, in case of
marriage. It is evident, therefore, that if both girls had married, this beauty would have had a
mere pittance, while even one of them would cripple him to a very serious extent. My morning's work
has not been wasted, since it has proved that he has the very strongest motives for standing in the
way of anything of the sort. And now, Watson, this is too serious for dawdling, especially as the
old man is aware that we are interesting ourselves in his affairs; so if you are ready, we shall
call a cab and drive to Waterloo. I should be very much obliged if you would slip your revolver into
your pocket. An Eley's No. 2 is an excellent argument with gentlemen who can twist steel pokers into
knots. That and a tooth-brush are, I think, all that we need."

At Waterloo we were fortunate in catching a train for Leatherhead, where we hired a trap at the
station inn and drove for four or five miles through the lovely Surrey lanes. It was a perfect day,
with a bright sun and a few fleecy clouds in the heavens. The trees and wayside hedges were just
throwing out their first green shoots, and the air was full of the pleasant smell of the moist
earth. To me at least there was a strange contrast between the sweet promise of the spring and this
sinister quest upon which we were engaged. My companion sat in the front of the trap, his arms
folded, his hat pulled down over his eyes, and his chin sunk upon his breast, buried in the deepest
thought. Suddenly, however, he started, tapped me on the shoulder, and pointed over the meadows.

"Look there!" said he.

A heavily timbered park stretched up in a gentle slope, thickening into a grove at the highest
point. From amid the branches there jutted out the grey gables and high roof-tree of a very old
mansion.