46 Corbet Close
"My dear fellow, is it possible you do not see how strongly it bears upon the case?"

"I cannot say that I do unless it were that he wished to be able to deny his signature if an action
for breach of promise were instituted."

"No, that was not the point. However, I shall write two letters, which should settle the matter. One
is to a firm in the City, the other is to the young lady's stepfather, Mr. Windibank, asking him
whether he could meet us here at six o'clock tomorrow evening. It is just as well that we should do
business with the male relatives. And now, Doctor, we can do nothing until the answers to those
letters come, so we may put our little problem upon the shelf for the interim."