51 Down Close
But I saw nothing. At the moment when Holmes struck the light I heard a low, clear whistle, but the
sudden glare flashing into my weary eyes made it impossible for me to tell what it was at which my
friend lashed so savagely. I could, however, see that his face was deadly pale and filled with
horror and loathing. He had ceased to strike and was gazing up at the ventilator when suddenly there
broke from the silence of the night the most horrible cry to which I have ever listened. It swelled
up louder and louder, a hoarse yell of pain and fear and anger all mingled in the one dreadful
shriek. They say that away down in the village, and even in the distant parsonage, that cry raised
the sleepers from their beds. It struck cold to our hearts, and I stood gazing at Holmes, and he at
me, until the last echoes of it had died away into the silence from which it rose.

"What can it mean?" I gasped.

"It means that it is all over," Holmes answered. "And perhaps, after all, it is for the best. Take
your pistol, and we will enter Dr. Roylott's room."