131 Wantage Close
"And how did you verify them?"

"Having once spotted my man, it was easy to get corroboration. I knew the firm for which this man
worked. Having taken the printed description, I eliminated everything from it which could be the
result of a disguise--the whiskers, the glasses, the voice, and I sent it to the firm, with a
request that they would inform me whether it answered to the description of any of their travellers.
I had already noticed the peculiarities of the typewriter, and I wrote to the man himself at his
business address asking him if he would come here. As I expected, his reply was typewritten and
revealed the same trivial but characteristic defects. The same post brought me a letter from
Westhouse & Marbank, of Fenchurch Street, to say that the description tallied in every respect with
that of their employee, James Windibank. Voilà tout!"

"And Miss Sutherland?"

"If I tell her she will not believe me. You may remember the old Persian saying, 'There is danger
for him who taketh the tiger cub, and danger also for whoso snatches a delusion from a woman.' There
is as much sense in Hafiz as in Horace, and as much knowledge of the world."