72 Sunningdale Drive
"I make nothing of it," I answered frankly. "It is a most mysterious business."

"As a rule," said Holmes, "the more bizarre a thing is the less mysterious it proves to be. It is
your commonplace, featureless crimes which are really puzzling, just as a commonplace face is the
most difficult to identify. But I must be prompt over this matter."

"What are you going to do, then?" I asked.

"To smoke," he answered. "It is quite a three pipe problem, and I beg that you won't speak to me for
fifty minutes." He curled himself up in his chair, with his thin knees drawn up to his hawk-like
nose, and there he sat with his eyes closed and his black clay pipe thrusting out like the bill of
some strange bird. I had come to the conclusion that he had dropped asleep, and indeed was nodding
myself, when he suddenly sprang out of his chair with the gesture of a man who has made up his mind
and put his pipe down upon the mantelpiece.

"Sarasate plays at the St. James's Hall this afternoon," he remarked. "What do you think, Watson?
Could your patients spare you for a few hours?"